Handle heterogeneous membranes
The -a/--atom option can now take several arguments to handle
heterogeneous membranes. The new format is "resname:atom_name". The
residue name can be omitted, then the format is just "atom_name".

This new feature does not break previous behavior.

The changes are undertested. The test suite still needs to be adapted so as
the error messages.