A file was not closed properly
When a file descriptor is used anonymously--i.e. it is not assigned to a
variable with a name--one can expect the file to be closed by the
garbage collector. This behavior seems to be a misconception of mine;
python 3.3 display a warning saying that the file is not closed
properly.

In this commit I add a with statement instead of an anonymous open so
the file is closed as it should be.