Move the ndx output in main
As mentioned in issue #20, the ndx output should not be in the
split_leaflets function. If it includes the ndx output, the
split_leaflets function as two purposes instead of one. Also,
split_leaflets is the closest we have to a entry point if splitleafs is
used as a module. Using the function with the ndx output in it results
in a very long and most likely undesirable print.