First step at refactoring splitleafs as a module
I previously updated splitleafs so it can be used as a python module.
Splitleafs still is a one-file script. Being so, building a proper API
for splitleafs is difficult and there is no code sharing with other
project. The file parsing, for instance, could be a separate module as I
copy-paste it between several scripts.

The refactoring will happen is several steps. This commit is the first
one. I created a splitleafs module as a directory with a __init__.py
file in it, and a command line tool called splitleafs.py. No code
changed so far. The repo is ugly at this stage but the tests still pass.

Next step will be to split the module in several files with everything
accessible the same way as it is now. Then an API will be properly
defined and documented.