Experiment with a new module entry point
This commit introduce structure.read which is a wrapper around
guess_format, read_gro, and read_pdb. This function aims at being a more
intelligent entry point for the splitleafs module.

The read function is now used in splitleafs.py. Hence, the format
guessing is moved out of the main function. If the main function is
still too long (see issue #10), it is reduced a bit.