Move the stat display from main to output
The stat display was the last part of non-cli logics in the main
function. It now exists as a function in the output module, which makes
more sense.

There is nothing left in the main function other than running the basic
functions and handle the exception they can raise. Even if the main
function still looks long, I think I can close issue #10.

Close #10.